add platform implementations
